Move the broadcast action to the generic tx status component
