+++ src/components/screens/reclaimBalance/modal/status/index.js
@@ -2,7 +2,6 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { withTranslation } from 'react-i18next';
 import { withRouter } from 'react-router';
-import { resetTransactionResult, transactionBroadcasted } from '@actions';
 import Status from './status';
 
 const mapStateToProps = state => ({
@@ -11,13 +10,8 @@ const mapStateToProps = state => ({
   network: state.network,
 });
 
-const mapDispatchToProps = {
-  transactionBroadcasted,
-  resetTransactionResult,
-};
-
 export default compose(
   withRouter,
-  connect(mapStateToProps, mapDispatchToProps),
+  connect(mapStateToProps),
   withTranslation(),
 )(Status);
