+++ src/components/screens/registerDelegate/status/index.js
@@ -1,7 +1,6 @@
 /* istanbul ignore file */
 import { connect } from 'react-redux';
 import { withTranslation } from 'react-i18next';
-import { transactionBroadcasted, resetTransactionResult } from '@actions';
 import { getActiveTokenAccount } from '@utils/account';
 import Status from './status';
 
@@ -11,12 +10,6 @@ const mapStateToProps = state => ({
   network: state.network,
 });
 
-const mapDispatchToProps = {
-  transactionBroadcasted,
-  resetTransactionResult,
-};
-
 export default connect(
   mapStateToProps,
-  mapDispatchToProps,
 )(withTranslation()(Status));
