+++ src/components/screens/registerDelegate/status/status.js
@@ -1,28 +1,15 @@
 import React, { useState, useEffect } from 'react';
-import { TransactionResult } from '@shared/transactionResult';
-import { isEmpty } from '@utils/helpers';
+import TransactionResult from '@shared/transactionResult';
 import DelegateAnimation from '../animations/delegateAnimation';
 import statusMessages from './statusMessages';
 import styles from './status.css';
 
 const Status = ({
   transactions, account, t,
-  transactionBroadcasted, resetTransactionResult,
 }) => {
   const [status, setStatus] = useState('pending');
   const isConfirmed = !!account.dpos.delegate?.username;
   const template = statusMessages(t)[status];
-
-  useEffect(() => {
-    if (!isEmpty(transactions.signedTransaction) && !transactions.txSignatureError) {
-      /**
-       * Broadcast the successfully signed tx
-       */
-      transactionBroadcasted(transactions.signedTransaction);
-    }
-
-    return resetTransactionResult;
-  }, []);
 
   useEffect(() => {
     if (transactions.txBroadcastError && status === 'pending') {
