+++ src/components/screens/votingQueue/result/index.js
@@ -4,7 +4,6 @@ import { compose } from 'redux';
 import { withRouter } from 'react-router';
 import { withTranslation } from 'react-i18next';
 import { getActiveTokenAccount } from '@utils/account';
-import { transactionBroadcasted, resetTransactionResult } from '@actions';
 import ResultComponent from './result';
 
 const mapStateToProps = state => ({
@@ -12,13 +11,8 @@ const mapStateToProps = state => ({
   transactions: state.transactions,
 });
 
-const mapDispatchToProps = {
-  resetTransactionResult,
-  transactionBroadcasted,
-};
-
 export default compose(
   withRouter,
-  connect(mapStateToProps, mapDispatchToProps),
+  connect(mapStateToProps),
   withTranslation(),
 )(ResultComponent);
