+++ src/components/screens/votingQueue/result/result.js
@@ -1,33 +1,20 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import { routes } from '@constants';
-import { isEmpty } from '@utils/helpers';
 import Box from '@toolbox/box';
 import BoxFooter from '@toolbox/box/footer';
 import { PrimaryButton } from '@toolbox/buttons';
-import { TransactionResult, getBroadcastStatus } from '@shared/transactionResult';
+import TransactionResult, { getBroadcastStatus } from '@shared/transactionResult';
 import ToggleIcon from '../toggleIcon';
 import statusMessages from './statusMessages';
 import styles from './styles.css';
 
 const Result = ({
   t, history, transactions, statusInfo,
-  resetTransactionResult, transactionBroadcasted,
 }) => {
   const closeModal = () => {
     history.push(routes.wallet.path);
   };
-
-  useEffect(() => {
-    if (!isEmpty(transactions.signedTransaction) && !transactions.txSignatureError) {
-      /**
-       * Broadcast the successfully signed tx
-       */
-      transactionBroadcasted(transactions.signedTransaction);
-    }
-
-    return resetTransactionResult;
-  }, []);
 
   const status = getBroadcastStatus(transactions, false); // @todo handle HW errors by #3661
   const template = statusMessages(t, statusInfo)[status.code];
